// src/common/enum/module-enum.ts
export enum Modules {
  COMMANDES = 'commandes',
  DASHBOARD = 'dashboard',
  MENU = 'menu',
  PROMOTIONS = 'promotions',
  CLIENTS = 'clients',
  CATEGORIES = 'categories',
  PLATS = 'plats',
  CHIFFRE_AFFAIRES = 'chiffre_affaires',
  ALL = 'all',
  MESSAGES = 'messages',
  RESTAURANT='restaurants',
  USERS='users',
  AUTH='auth',
  DISHES='dishes',
  ADDRESSES='addresses',
  FAVORITES='favorites',
  COMMENTS='comments',
  PAIEMENTS='paiements',
  NOTIFICATIONS='notifications',
  LOYALITY='loyality',
  STATISTICS='statistics',
  EMAIL='email',
  CONVERSATIONS='conversations',
  
}